updates file org and dependency for react-use-dimensions
